clipmenud: correct cast to size_t

+++ src/clipmenud.c
@@ -191,8 +191,7 @@ static void maybe_trim(void) {
     uint64_t cur_clips;
     expect(cs_len(&cs, &cur_clips) == 0);
     if ((int)cur_clips > cfg.max_clips_batch) {
-        expect(cs_trim(&cs, CS_ITER_NEWEST_FIRST, (uint64_t)cfg.max_clips) ==
-               0);
+        expect(cs_trim(&cs, CS_ITER_NEWEST_FIRST, (size_t)cfg.max_clips) == 0);
     }
 }
